+++ ErsatzTV.Application/Configuration/ThemeMode.cs
@@ -1,0 +1,8 @@
+namespace ErsatzTV.Application.Configuration;
+
+public enum ThemeMode
+{
+    Dark = 0,
+    Light = 1,
+    System = 2
+}