+++ ErsatzTV.Application/Configuration/UiSettingsViewModel.cs
@@ -2,7 +2,7 @@ namespace ErsatzTV.Application.Configuration;
 
 public class UiSettingsViewModel
 {
-    public bool IsDarkMode { get; set; }
+    public ThemeMode ThemeMode { get; set; }
 
     public string Language { get; set; }
 }